+++ core/shared-ui/src/desktopMain/kotlin/com/neoutils/neoregex/core/sharedui/component/NeoWindow.kt
@@ -40,8 +40,8 @@ import com.jetbrains.JBR
 import com.neoutils.neoregex.core.resources.Res
 import com.neoutils.neoregex.core.resources.app_name
 import com.neoutils.neoregex.core.resources.flavicon
-import com.neoutils.neoregex.core.sharedui.remember.NeoWindowState
-import com.neoutils.neoregex.core.sharedui.remember.rememberNeoWindowState
+import com.neoutils.neoregex.core.sharedui.extension.NeoWindowState
+import com.neoutils.neoregex.core.sharedui.extension.rememberNeoWindowState
 import com.neoutils.neoregex.core.sharedui.util.NeoRegexWindowExceptionHandlerFactory
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
