+++ core/shared-ui/src/desktopMain/kotlin/com/neoutils/neoregex/core/sharedui/extension/FrameWindowScope.kt
@@ -1,7 +1,7 @@
 /*
  * NeoRegex.
  *
- * Copyright (C) 2024 Irineu A. Silva.
+ * Copyright (C) 2025 Irineu A. Silva.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,12 +16,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neoutils.neoregex.core.sharedui.remember
+package com.neoutils.neoregex.core.sharedui.extension
 
 import androidx.compose.runtime.*
+import androidx.compose.ui.awt.ComposeWindow
 import androidx.compose.ui.window.FrameWindowScope
 import java.awt.event.WindowAdapter
 import java.awt.event.WindowEvent
+import java.awt.event.WindowStateListener
 
 enum class WindowFocus {
     FOCUSED,
@@ -61,4 +63,47 @@ fun FrameWindowScope.rememberWindowFocus(): WindowFocus {
     }
 
     return focus
+}
+
+enum class NeoWindowState {
+    FLOATING,
+    MAXIMIZED,
+    FULLSCREEN,
+    MINIMIZED,
+    PINNED;
+
+    companion object {
+        fun of(state: ComposeWindow): NeoWindowState {
+
+            return when {
+                state.isFloating -> FLOATING
+                state.isFullMaximized -> MAXIMIZED
+                state.isHalfMaximized -> PINNED
+                state.isFullscreen -> FULLSCREEN
+                state.isMinimized -> MINIMIZED
+                else -> error("Unknown window state")
+            }
+        }
+    }
+}
+
+@Composable
+fun FrameWindowScope.rememberNeoWindowState(): NeoWindowState {
+
+    var state by remember { mutableStateOf(NeoWindowState.of(window)) }
+
+    DisposableEffect(window) {
+
+        val stateListener = WindowStateListener {
+            state = NeoWindowState.of(window)
+        }
+
+        window.addWindowStateListener(stateListener)
+
+        onDispose {
+            window.removeWindowStateListener(stateListener)
+        }
+    }
+
+    return state
 }
